requirements.txt
wtforms
m3u8
requests
websockets
numpy
click
flask
celery
flower
Flask-SQLAlchemy
flask-admin
flask-babelex
opencv-python
redis
configparser
# apt-get install libmysqlclient-dev
mysqlclient
shapely
# apt-get install ffmpeg
ffmpeg-python